refactor: verfificationForm, Button component
