+++ src/app/signUp/SignUpForm.tsx
@@ -115,6 +115,7 @@ export default function SignUpForm() {
 
 
 
+    //RENDER
     return (
         <div className="flex items-center justify-center w-full min-h-screen">
