+++ src/components/Button.tsx
@@ -16,13 +16,38 @@ const defaultStyles: React.CSSProperties = {
   fontSize: "16px",
   cursor: "pointer",
   padding: "12px",
-  
+  transition: "all 0.2s ease-in-out",
 };
 
-const Button: React.FC<ButtonProps> = ({ background = "#786599", style, children, ...props }) => {
+const Button: React.FC<ButtonProps> = ({ background = "#786599", style, children, disabled, ...props }) => {
+  const buttonStyles: React.CSSProperties = {
+    ...defaultStyles,
+    background: disabled ? "#cccccc" : background,
+    cursor: disabled ? "not-allowed" : "pointer",
+    opacity: disabled ? 0.6 : 1,
+    ...style,
+  };
+
+  const handleMouseEnter = (e: React.MouseEvent<HTMLButtonElement>) => {
+    if (!disabled && e.currentTarget) {
+      e.currentTarget.style.opacity = "0.8"; // Ciemniejszy efekt przez zmniejszenie opacity
+      e.currentTarget.style.transform = "translateY(-1px)";
+    }
+  };
+
+  const handleMouseLeave = (e: React.MouseEvent<HTMLButtonElement>) => {
+    if (!disabled && e.currentTarget) {
+      e.currentTarget.style.opacity = "1";
+      e.currentTarget.style.transform = "translateY(0)";
+    }
+  };
+
   return (
     <button
-      style={{ ...defaultStyles, background, ...style }}
+      style={buttonStyles}
+      onMouseEnter={handleMouseEnter}
+      onMouseLeave={handleMouseLeave}
+      disabled={disabled}
       {...props}
     >
       {children}
