+++ src/styles/globals.css
@@ -115,10 +115,10 @@ html, body {
 }
 
 
-button:hover {
+/* button:hover {
   filter: brightness(0.8);
   transition: filter 0.2s;
-}
+} */
